Add files via upload
This Java program is a simple Notes Manager that lets users add and view notes using File I/O. It displays a menu where users can choose to add a note, view all saved notes, or exit. When the user adds a note, the program takes input from the console and stores it in a text file (notes.txt) using FileWriter in append mode, so previous notes are not erased. When the user chooses to view notes, the program reads the file line by line using BufferedReader and displays each note on the screen. It also handles errors properly: if the file doesn’t exist, it tells the user that no notes are found, and if any I/O error occurs, it shows an error message. The program runs in a loop until the user selects exit, ensuring continuous interaction.