Add files via upload
This Java code implements a To-Do List application using Swing components. It creates a main window (JFrame) featuring an input field and an "Add" button at the top, a scrollable list (JList) in the center to display tasks, and "Delete Selected" and "Clear All" buttons at the bottom. Tasks are managed using a DefaultListModel. The application is interactive: users can add tasks by typing and pressing Enter or clicking "Add," delete selected tasks using a button or the Delete key, clear the entire list, and double-click a task to edit it via an input dialog. All GUI operations are safely executed on the Event Dispatch Thread (EDT).